+++ src/app/router/routesConfig.tsx
@@ -1,5 +1,4 @@
 import MainPage from '../../pages/MainPage';
-import HeaderCopy from '../../widgets/HeaderCopy';
 import NotFoundPage from '../../widgets/NotFoundPage/index';
 
 import { IRoute } from './type';
@@ -8,15 +7,6 @@ export const routesConfig: IRoute[] = [
   {
     path: '/',
     element: <MainPage />,
-    auth: false
-  },
-  {
-    path: '/test',
-    element: (
-      <div style={{ margin: '20px auto' }}>
-        <HeaderCopy auth={'student'} />
-      </div>
-    ),
     auth: false
   },
   {
