+++ src/widgets/Header/AuthHeader/index.tsx
