+++ src/widgets/Header/types.ts
@@ -1,9 +1,23 @@
-export type TAuth = 'unauth' | 'student' | 'teacher';
+export type TUserRole = 'unauth' | 'student' | 'tutor';
 
 export interface HeaderProps {
-  auth: TAuth;
+  auth: TUserRole;
 }
 
 export interface CommonHeaderProps {
+  role: Exclude<TUserRole, 'unauth'>;
   onLogout: () => void;
+}
+
+export interface NavItem {
+  text: string;
+  path?: string;
+  icon?: string;
+  onClick?: () => void;
+}
+
+export interface AuthHeaderProps {
+  navItems: NavItem[];
+  avatarMenuItems: NavItem[];
+  avatarImage?: string;
 }
