+++ src/widgets/Header/useHeaderConfig.ts
