--- src/widgets/HeaderCopy/index.module.scss
@@ -1,106 +1,0 @@
-.header {
-  display: flex;
-  justify-content: space-between;
-  align-items: center;
-  border-radius: var(--border-radius-blocks);
-  padding: 26px 32px;
-  background-color: var(--clr-main);
-  max-inline-size: 1200px;
-
-  &__logo {
-    cursor: pointer;
-  }
-
-  &__menu {
-    display: flex;
-    gap: 48px;
-  }
-
-  &__button {
-    font-family: var(--ffm-default, 'Montserrat', courier, monospace);
-    font-weight: 400;
-    font-size: 16px;
-    line-height: 1;
-    border: 2px solid transparent;
-    background: none;
-    transition: border-color 0.3s;
-    cursor: pointer;
-
-    &:hover {
-      font-weight: 600;
-      font-size: 16px;
-      line-height: 20px;
-      border-bottom: 2px solid
-        var(--clr-primary-default, var(--clr-primary-default));
-      color: var(--clr-primary-default);
-    }
-  }
-
-  &__svg {
-    margin-inline-end: 10px;
-  }
-
-  &__avatar {
-    position: relative;
-    cursor: pointer;
-    padding-right: 40px;
-
-    &__chevron {
-      content: '';
-      position: absolute;
-      top: 20%;
-      right: 8px;
-      width: 24px;
-      height: 24px;
-      border: none;
-      background: none;
-      background-image: url('../../assets/images/arrow.svg');
-      transition: transform 0.3s ease;
-
-      &--active {
-        transform: rotate(180deg);
-      }
-    }
-  }
-
-  &__submenu {
-    position: absolute;
-    top: 51px;
-    right: 5px;
-    visibility: hidden;
-    opacity: 0;
-    transition: all 0.3s ease;
-    transform: translateY(-11px);
-    background-color: var(--clr-main);
-    border-radius: var(--border-radius-blocks);
-
-    &__list {
-      display: flex;
-      flex-direction: column;
-      gap: 7px;
-      font-weight: 500;
-
-      &__item {
-        display: flex;
-        gap: 9px;
-        align-items: center;
-        padding: 8px;
-        font-size: 14px;
-        border: none;
-        background: none;
-        inline-size: 191px;
-      }
-    }
-
-    &--active {
-      visibility: visible;
-      opacity: 1;
-      transform: translateY(0);
-    }
-  }
-}
-
-.telegram {
-  padding-inline: 17px !important;
-  padding-block: 19px;
-}
